+++ src/input/keyboard.rs
@@ -1,6 +1,10 @@
 use crate::Context;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 #[allow(missing_docs)]
 /// A key on a keyboard.
 pub enum Key {