+++ src/time.rs
@@ -8,6 +8,10 @@ use crate::Context;
 
 /// The different timestep modes that a game can have.
 #[derive(Debug, Copy, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 pub enum Timestep {
     /// In fixed timestep mode, updates will happen at a consistent rate (the `f64` value in the enum
     /// variant representing the number of times per second), while rendering will happen as fast as